mail/requirements.txt
google-auth
google-auth-oauthlib>=0.5.2
google-api-python-client
python-dotenv